add logic to perform RTT Ranging based on user settings

+++ app/src/main/java/com/davidperez/tfgwifirtt/model/UserSettings.kt
@@ -3,7 +3,7 @@ package com.davidperez.tfgwifirtt.model
 data class UserSettings(
     val showOnlyRttCompatibleAps: Boolean = false,
     val performContinuousRttRanging: Boolean = false,
-    val rttPeriod: Int = 10,
-    val rttInterval: Int = 500,
+    val rttPeriod: Long = 10,
+    val rttInterval: Long = 500,
     val saveRttResults: Boolean = true
 )
